refactor test class for security and delete jwt auth filter test
